added image uploading

+++ src/express/badges/delete.ts
@@ -1,7 +1,14 @@
 import WrappedRouter from "@lib/utils/WrappedRouter";
 import { Badge } from "@mongoose/schemas/Badge";
 import { BadgesDelete } from "@shared/ts/api/badges";
-import { error, notFoundError, ok, unauthorizedError, validationError } from "@shared/utils/api";
+import {
+  error,
+  forbiddenError,
+  notFoundError,
+  ok,
+  unauthorizedError,
+  validationError,
+} from "@shared/utils/api";
 import badgeValidators from "@shared/validators/badgeValidators";
 import { parseValidators } from "@shared/utils/generic";
 
@@ -27,6 +34,9 @@ router.delete<BadgesDelete>("/:badgeId", async (req, res) => {
     ]);
 
     if (!badge) return notFoundError("No badge exists with that id")(res);
+
+    if (!badge.createdBy._id.equals(req.user._id))
+      return forbiddenError("You are not the creator of this badge")(res);
 
     await badge.delete();
